requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==7.0.1
Pillow==8.1.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
PyJWT==2.0.1